+++ appseed/ca/sockets/timer/sockets_EventHandler.cpp
@@ -77,7 +77,7 @@ namespace sockets
             diff = 1;
          }
          tv -> tv_sec = static_cast<long>(diff / 1000000);
-         tv -> tv_usec = static_cast<long>(diff % 1000000);
+         tv -> tv_usec = static_cast<int>(diff % 1000000);
          return true;
       }
       return false;
